Use '.remote_field' and '.model' in preference to '.rel' and '.to' when inspecting model fields.
