+++ rest_framework/utils/model_meta.py
@@ -5,15 +5,9 @@ relationships and their associated metadata.
 
 Usage: `get_field_info(model)` returns a `FieldInfo` instance.
 """
-import inspect
 from collections import OrderedDict, namedtuple
 
-from django.apps import apps
-from django.core.exceptions import ImproperlyConfigured
-from django.db import models
-from django.utils import six
-
-from rest_framework.compat import get_remote_field
+from rest_framework.compat import get_related_model, get_remote_field
 
 FieldInfo = namedtuple('FieldResult', [
     'pk',  # Model field instance
@@ -31,30 +25,6 @@ RelationInfo = namedtuple('RelationInfo', [
     'to_field',
     'has_through_model'
 ])
-
-
-def _resolve_model(obj):
-    """
-    Resolve supplied `obj` to a Django model class.
-
-    `obj` must be a Django model class itself, or a string
-    representation of one.  Useful in situations like GH #1225 where
-    Django may not have resolved a string-based reference to a model in
-    another model's foreign key definition.
-
-    String representations should have the format:
-        'appname.ModelName'
-    """
-    if isinstance(obj, six.string_types) and len(obj.split('.')) == 2:
-        app_name, model_name = obj.split('.')
-        resolved_model = apps.get_model(app_name, model_name)
-        if resolved_model is None:
-            msg = "Django did not return a model for {0}.{1}"
-            raise ImproperlyConfigured(msg.format(app_name, model_name))
-        return resolved_model
-    elif inspect.isclass(obj) and issubclass(obj, models.Model):
-        return obj
-    raise ValueError("{0} is not a Django model".format(obj))
 
 
 def get_field_info(model):
@@ -82,7 +52,7 @@ def _get_pk(opts):
 
     while rel and rel.parent_link:
         # If model is a child via multi-table inheritance, use parent's pk.
-        pk = rel.to._meta.pk
+        pk = get_related_model(pk)._meta.pk
         rel = get_remote_field(pk)
 
     return pk
@@ -108,7 +78,7 @@ def _get_forward_relationships(opts):
     for field in [field for field in opts.fields if field.serialize and get_remote_field(field)]:
         forward_relations[field.name] = RelationInfo(
             model_field=field,
-            related_model=_resolve_model(get_remote_field(field).to),
+            related_model=get_related_model(field),
             to_many=False,
             to_field=_get_to_field(field),
             has_through_model=False
@@ -118,7 +88,7 @@ def _get_forward_relationships(opts):
     for field in [field for field in opts.many_to_many if field.serialize]:
         forward_relations[field.name] = RelationInfo(
             model_field=field,
-            related_model=_resolve_model(get_remote_field(field).to),
+            related_model=get_related_model(field),
             to_many=True,
             # manytomany do not have to_fields
             to_field=None,
