+++ tests/test_utils.py
@@ -6,8 +6,8 @@ from django.test import TestCase, override_settings
 from django.utils import six
 
 import rest_framework.utils.model_meta
+from rest_framework.compat import _resolve_model
 from rest_framework.utils.breadcrumbs import get_breadcrumbs
-from rest_framework.utils.model_meta import _resolve_model
 from rest_framework.views import APIView
 from tests.models import BasicModel
 
@@ -166,16 +166,16 @@ class ResolveModelWithPatchedDjangoTests(TestCase):
 
     def setUp(self):
         """Monkeypatch get_model."""
-        self.get_model = rest_framework.utils.model_meta.apps.get_model
+        self.get_model = rest_framework.compat.apps.get_model
 
         def get_model(app_label, model_name):
             return None
 
-        rest_framework.utils.model_meta.apps.get_model = get_model
+        rest_framework.compat.apps.get_model = get_model
 
     def tearDown(self):
         """Revert monkeypatching."""
-        rest_framework.utils.model_meta.apps.get_model = self.get_model
+        rest_framework.compat.apps.get_model = self.get_model
 
     def test_blows_up_if_model_does_not_resolve(self):
         with self.assertRaises(ImproperlyConfigured):
